refactor_test добавление запуска через selenium grid, рефактор base_page, добавление requirments
